+++ packages/fusion/src/check.rs
@@ -42,60 +42,22 @@ pub mod tokio {
     pub use crate::tokio::check::CheckAsyncExt;
 }
 
-/// Error type of the result from the check process.
+/// Check process missing chunks error.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct MissingChunks {
+    // pub missing: Vec<usize>,
+    pub missing: Vec<usize>,
+}
+
+/// Check process size mismatch error.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CheckResultErrorType {
-    /// Some of the chunks are missing to merge the file.
-    Missing,
-    /// The actual file size is not equal the input file size.
-    Size,
+pub struct SizeMismatch {
+    pub expected: usize,
+    pub actual: usize,
 }
 
-impl CheckResultErrorType {
-    /// Get the error type from code.
-    pub fn from_code<C: AsRef<str>>(code: C) -> Option<Self> {
-        match code.as_ref() {
-            | "missing" => Some(Self::Missing),
-            | "size" => Some(Self::Size),
-            | _ => None,
-        }
-    }
-
-    /// Get the code of the error type as `&str`.
-    pub fn as_code(&self) -> &str {
-        match self {
-            | Self::Missing => "missing",
-            | Self::Size => "size",
-        }
-    }
-
-    /// Get the code of the error type as `String`.
-    pub fn to_code(&self) -> String {
-        self.as_code().to_string()
-    }
-}
-
-/// Error of the result from the check process.
-#[derive(Debug, Clone)]
-pub struct CheckResultError {
-    /// Type of error of the check.
-    pub error_type: CheckResultErrorType,
-    /// Error message of the check.
-    pub message: String,
-    /// Missing chunk(s) to merge the file.
-    pub missing: Option<Vec<usize>>,
-}
-
-/// Result of the check process.
-#[derive(Debug, Clone)]
-pub struct CheckResult {
-    /// Successful / Failed check.
-    pub success: bool,
-    /// Error details of the check.
-    pub error: Option<CheckResultError>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// Check process error enum.
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum CheckError {
     InDirNotFound,
     InDirNotDir,
@@ -104,6 +66,8 @@ pub enum CheckError {
     InFileNotRead,
     FileSizeNotSet,
     TotalChunksNotSet,
+    MissingChunks(MissingChunks),
+    SizeMismatch(SizeMismatch),
 }
 
 impl CheckError {
@@ -117,6 +81,8 @@ impl CheckError {
             | Self::InFileNotRead => "in_file_not_read",
             | Self::FileSizeNotSet => "file_size_not_set",
             | Self::TotalChunksNotSet => "total_chunks_not_set",
+            | Self::MissingChunks(_) => "missing_chunks",
+            | Self::SizeMismatch(_) => "size_mismatch",
         }
     }
 
@@ -135,6 +101,12 @@ impl CheckError {
             | Self::InFileNotRead => "The input file could not be read.",
             | Self::FileSizeNotSet => "The `file_size` is not set.",
             | Self::TotalChunksNotSet => "The `total_chunks` is not set.",
+            | Self::MissingChunks(_) => {
+                "Some of the chunks are missing to merge the file."
+            },
+            | Self::SizeMismatch(_) => {
+                "The actual file size is not equal the input file size."
+            },
         }
     }
 
@@ -155,14 +127,26 @@ impl CheckError {
 /// ```no_run
 /// use std::path::PathBuf;
 ///
-/// use filerune_fusion::check::{Check, CheckResult};
+/// use filerune_fusion::check::{Check, CheckError};
 ///
-/// let result: CheckResult = Check::new()
+/// if let Err(error) = Check::new()
 ///     .in_dir(PathBuf::from("path").join("to").join("dir"))
 ///     .file_size(0) // result from split function...
 ///     .total_chunks(0) // result from split function...
 ///     .run()
-///     .unwrap();
+/// {
+///     match error {
+///         | CheckError::MissingChunks(err) => {
+///             // ...
+///         },
+///         | CheckError::SizeMismatch(err) => {
+///             // ...
+///         },
+///         | _ => {
+///             // ...
+///         },
+///     }
+/// };
 /// ```
 #[derive(Debug, Clone)]
 pub struct Check {
@@ -210,7 +194,7 @@ impl Check {
     }
 
     /// Run the check process.
-    pub fn run(&self) -> Result<CheckResult, CheckError> {
+    pub fn run(&self) -> Result<bool, CheckError> {
         let in_dir: &Path = match self.in_dir {
             | Some(ref p) => {
                 let p: &Path = p.as_ref();
@@ -263,30 +247,17 @@ impl Check {
         }
 
         if !missing.is_empty() {
-            return Ok(CheckResult {
-                success: false,
-                error: Some(CheckResultError {
-                    error_type: CheckResultErrorType::Missing,
-                    message: "Missing chunk(s)".to_string(),
-                    missing: Some(missing),
-                }),
-            });
+            return Err(CheckError::MissingChunks(MissingChunks { missing }));
         }
 
-        if actual_size != file_size {
-            return Ok(CheckResult {
-                success: false,
-                error: Some(CheckResultError {
-                    error_type: CheckResultErrorType::Size,
-                    message:
-                        "the size of chunks is not equal to file_size parameter"
-                            .to_string(),
-                    missing: None,
-                }),
-            });
+        if file_size != actual_size {
+            return Err(CheckError::SizeMismatch(SizeMismatch {
+                expected: file_size,
+                actual: actual_size,
+            }));
         }
 
-        Ok(CheckResult { success: true, error: None })
+        Ok(true)
     }
 }